configure chains
