+++ examples/local-explorer/mprocs.yaml
@@ -18,4 +18,4 @@ procs:
       SQLITE_FILENAME: "indexer.db"
   explorer:
     cwd: packages/contracts
-    shell: pnpm explorer --dev --chainId 17069 --worldAddress 0x9ca094a38027bb331656e4f1f6ed40c9982dbe70
+    shell: pnpm explorer --dev
