+++ packages/explorer/src/common.ts
@@ -1,0 +1,9 @@
+import { Chain, anvil, garnet, redstone } from "viem/chains";
+
+const chains = {
+  [anvil.id]: anvil,
+  [redstone.id]: redstone,
+  [garnet.id]: garnet,
+};
+export const chainId = Number(process.env.NEXT_PUBLIC_CHAIN_ID || anvil.id);
+export const chain = chains[chainId as keyof typeof chains] as Chain;
